requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1

# Env
python-dotenv==1.0.1
pydantic-settings==2.3.4

# LangChain stack
langchain==0.2.11
langchain-community==0.2.10
langchain-groq==0.1.6
langgraph==0.0.69

# Embeddings
langchain-huggingface==0.0.3
sentence-transformers==2.7.0

# Vector DB
faiss-cpu==1.11.0

# Document loaders
pypdf==4.2.0
python-docx==1.1.2
docx2txt==0.9